refactor: Split http.py into modular directory structure for pyxhr integration
