+++ src/py/pyodide/http/pyfetch.py
@@ -1,3 +1,7 @@
+"""
+Async fetch API implementation for Pyodide.
+"""
+
 import json
 from asyncio import CancelledError
 from collections.abc import Awaitable, Callable
@@ -5,18 +9,16 @@ from functools import wraps
 from io import StringIO
 from typing import IO, Any, ParamSpec, TypeVar
 
-from ._package_loader import unpack_buffer
-from .ffi import IN_BROWSER, JsBuffer, JsException, JsFetchResponse, to_js
+from .._package_loader import unpack_buffer
+from ..ffi import IN_BROWSER, JsBuffer, JsException, JsFetchResponse, to_js
+from ._utils import _construct_abort_reason
+from .exceptions import AbortError, BodyUsedError, HttpStatusError
 
 if IN_BROWSER:
     try:
         from js import AbortController, AbortSignal, Object
         from js import fetch as _jsfetch
         from pyodide_js._api import abortSignalAny
-    except ImportError:
-        pass
-    try:
-        from js import XMLHttpRequest
     except ImportError:
         pass
 else:
@@ -28,102 +30,6 @@ else:
 
     class AbortSignal:  # type:ignore[no-redef]
         pass
-
-
-__all__ = [
-    "open_url",
-    "pyfetch",
-    "FetchResponse",
-    "HttpStatusError",
-    "BodyUsedError",
-    "AbortError",
-]
-
-
-class HttpStatusError(OSError):
-    """A subclass of :py:exc:`OSError` raised by :py:meth:`FetchResponse.raise_for_status`
-    if the response status is 4XX or 5XX.
-
-    Parameters
-    ----------
-    status :
-       The http status code of the request
-
-    status_text :
-       The http status text of the request
-
-    url :
-        The url that was requested
-    """
-
-    status: int
-    status_text: str
-    url: str
-
-    def __init__(self, status: int, status_text: str, url: str) -> None:
-        self.status = status
-        self.status_text = status_text
-        self.url = url
-        if 400 <= status < 500:
-            super().__init__(f"{status} Client Error: {status_text} for url: {url}")
-        elif 500 <= status < 600:
-            super().__init__(f"{status} Server Error: {status_text} for url: {url}")
-        else:
-            super().__init__(
-                f"{status} Invalid error code not between 400 and 599: {status_text} for url: {url}"
-            )
-
-    def __reduce__(self):
-        return (
-            self.__class__,
-            (self.status, self.status_text, self.url),
-            self.__dict__,
-        )
-
-
-class BodyUsedError(OSError):
-    def __init__(self, *args: Any) -> None:
-        super().__init__("Response body is already used")
-
-
-class AbortError(OSError):
-    def __init__(self, reason: JsException) -> None:
-        super().__init__(reason.message)
-
-
-def open_url(url: str) -> StringIO:
-    """Fetches a given URL synchronously.
-
-    The download of binary files is not supported. To download binary files use
-    :func:`pyodide.http.pyfetch` which is asynchronous.
-
-    It will not work in Node unless you include an polyfill for :js:class:`XMLHttpRequest`.
-
-    Parameters
-    ----------
-    url :
-       URL to fetch
-
-    Returns
-    -------
-        The contents of the URL.
-
-    Examples
-    --------
-    >>> None # doctest: +RUN_IN_PYODIDE
-    >>> import pytest; pytest.skip("TODO: Figure out how to skip this only in node")
-    >>> url = "https://cdn.jsdelivr.net/pyodide/v0.24.1/full/pyodide-lock.json"
-    >>> url_contents = open_url(url)
-    >>> import json
-    >>> result = json.load(url_contents)
-    >>> sorted(list(result["info"].items()))
-    [('arch', 'wasm32'), ('platform', 'emscripten_3_1_45'), ('python', '3.11.3'), ('version', '0.24.1')]
-    """
-
-    req = XMLHttpRequest.new()
-    req.open("GET", url, False)
-    req.send()
-    return StringIO(req.response)
 
 
 P = ParamSpec("P")
@@ -149,12 +55,6 @@ def _abort_on_cancel(method: Callable[P, Awaitable[T]]) -> Callable[P, Awaitable
             raise
 
     return wrapper
-
-
-def _construct_abort_reason(reason: Any) -> JsException | None:
-    if reason is None:
-        return None
-    return JsException("AbortError", reason)
 
 
 class FetchResponse:
